requirements.txt
fastapi
pydantic
youtube-transcript-api
uvicorn
requests
python-dotenv